+++ lib/plug_rest/conn.ex
@@ -23,6 +23,28 @@ defmodule PlugRest.Conn do
   @type header_value :: {String.t, map()}
 
   @doc """
+  Reads the dynamic segment values from a rest resource path
+
+  #Examples
+
+      iex > PlugRest.Conn.read_path_params(conn)
+      %{"id" => 123}
+
+  """
+  @spec read_path_params(conn, Keyword.t) :: %{binary => binary}
+  def read_path_params(conn, opts \\ [])
+
+  def read_path_params(%Plug.Conn{private: %{plug_rest_path_params: params}},
+                      _opts) do
+    params
+  end
+
+  def read_path_params(_conn, _opts) do
+    %{}
+  end
+
+
+  @doc """
   Parses request date header as Erlang date/time tuples
 
   Possible headers are: